LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-popup-element/popup-anchor-nesting.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: popup2.show is not a function. (In 'popup2.show()', 'popup2.show' is undefined)
Popup 1 Inside popup 2  This is popup 1 Popup 2

Harness Error (FAIL), message = Unhandled rejection: popup2.show is not a function. (In 'popup2.show()', 'popup2.show' is undefined)
